feat: Sprint 1 — KMP foundation, designsystem theme migration, composeApp skeleton
- Add Compose Multiplatform (1.7.3) and KMP navigation (2.8.0-alpha12) to version catalog
- Convert :core:designsystem to KMP with androidTarget + iOS targets
- Convert :core:domain and :core:data from jvm() to androidTarget()
- Bundle Anuphan + ADLaM Display fonts in composeResources/font/
- Migrate Color.kt, Type.kt, Theme.kt to commonMain with expect/actual for dynamic color
- Create :composeApp KMP skeleton module
- Add backward-compat theme wrappers in :app
