+++ build.gradle.kts
@@ -8,5 +8,6 @@ plugins {
     alias(libs.plugins.kotlin.compose) apply false
     alias(libs.plugins.kotlin.serialization) apply false
     alias(libs.plugins.ksp) apply false
+    alias(libs.plugins.compose.multiplatform) apply false
     id("com.google.gms.google-services") version "4.4.4" apply false
 }
