+++ core/data/build.gradle.kts
@@ -1,10 +1,11 @@
 plugins {
     alias(libs.plugins.kotlin.multiplatform)
     alias(libs.plugins.kotlin.serialization)
+    alias(libs.plugins.android.library)
 }
 
 kotlin {
-    jvm()
+    androidTarget()
     iosArm64()
     iosSimulatorArm64()
 
@@ -19,5 +20,19 @@ kotlin {
             implementation(libs.gitlive.firebase.auth)
             implementation(libs.gitlive.firebase.firestore)
         }
+    }
+}
+
+android {
+    namespace = "com.sedilant.yambol.core.data"
+    compileSdk = 36
+
+    defaultConfig {
+        minSdk = 24
+    }
+
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
     }
 }
