+++ core/designsystem/src/androidMain/kotlin/com/sedilant/yambol/core/designsystem/theme/PlatformTheme.android.kt
@@ -1,0 +1,18 @@
+package com.sedilant.yambol.core.designsystem.theme
+
+import android.os.Build
+import androidx.compose.material3.ColorScheme
+import androidx.compose.material3.dynamicDarkColorScheme
+import androidx.compose.material3.dynamicLightColorScheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.platform.LocalContext
+
+@Composable
+actual fun platformDynamicColorScheme(darkTheme: Boolean): ColorScheme? {
+    return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
+        val context = LocalContext.current
+        if (darkTheme) dynamicDarkColorScheme(context) else dynamicLightColorScheme(context)
+    } else {
+        null
+    }
+}
