+++ core/designsystem/src/commonMain/kotlin/com/sedilant/yambol/core/designsystem/theme/Color.kt
@@ -1,0 +1,219 @@
+package com.sedilant.yambol.core.designsystem.theme
+
+import androidx.compose.ui.graphics.Color
+val primaryLight = Color(0xFF39608F)
+val onPrimaryLight = Color(0xFFFFFFFF)
+val primaryContainerLight = Color(0xFFD3E4FF)
+val onPrimaryContainerLight = Color(0xFF1E4876)
+val secondaryLight = Color(0xFF8B4F25)
+val onSecondaryLight = Color(0xFFFFFFFF)
+val secondaryContainerLight = Color(0xFFFFDBC7)
+val onSecondaryContainerLight = Color(0xFF6E380F)
+val tertiaryLight = Color(0xFF86521A)
+val onTertiaryLight = Color(0xFFFFFFFF)
+val tertiaryContainerLight = Color(0xFFFFDCBF)
+val onTertiaryContainerLight = Color(0xFF6A3B02)
+val errorLight = Color(0xFFBA1A1A)
+val onErrorLight = Color(0xFFFFFFFF)
+val errorContainerLight = Color(0xFFFFDAD6)
+val onErrorContainerLight = Color(0xFF93000A)
+val backgroundLight = Color(0xFFF8F9FF)
+val onBackgroundLight = Color(0xFF191C20)
+val surfaceLight = Color(0xFFF5FAFB)
+val onSurfaceLight = Color(0xFF171D1E)
+val surfaceVariantLight = Color(0xFFDFE2EB)
+val onSurfaceVariantLight = Color(0xFF43474E)
+val outlineLight = Color(0xFF73777F)
+val outlineVariantLight = Color(0xFFC3C6CF)
+val scrimLight = Color(0xFF000000)
+val inverseSurfaceLight = Color(0xFF2B3133)
+val inverseOnSurfaceLight = Color(0xFFECF2F3)
+val inversePrimaryLight = Color(0xFFA3C9FE)
+val surfaceDimLight = Color(0xFFD5DBDC)
+val surfaceBrightLight = Color(0xFFF5FAFB)
+val surfaceContainerLowestLight = Color(0xFFFFFFFF)
+val surfaceContainerLowLight = Color(0xFFEFF5F6)
+val surfaceContainerLight = Color(0xFFE9EFF0)
+val surfaceContainerHighLight = Color(0xFFE3E9EA)
+val surfaceContainerHighestLight = Color(0xFFDEE3E5)
+
+val primaryLightMediumContrast = Color(0xFF043764)
+val onPrimaryLightMediumContrast = Color(0xFFFFFFFF)
+val primaryContainerLightMediumContrast = Color(0xFF496F9F)
+val onPrimaryContainerLightMediumContrast = Color(0xFFFFFFFF)
+val secondaryLightMediumContrast = Color(0xFF5A2801)
+val onSecondaryLightMediumContrast = Color(0xFFFFFFFF)
+val secondaryContainerLightMediumContrast = Color(0xFF9D5D32)
+val onSecondaryContainerLightMediumContrast = Color(0xFFFFFFFF)
+val tertiaryLightMediumContrast = Color(0xFF532D00)
+val onTertiaryLightMediumContrast = Color(0xFFFFFFFF)
+val tertiaryContainerLightMediumContrast = Color(0xFF986127)
+val onTertiaryContainerLightMediumContrast = Color(0xFFFFFFFF)
+val errorLightMediumContrast = Color(0xFF740006)
+val onErrorLightMediumContrast = Color(0xFFFFFFFF)
+val errorContainerLightMediumContrast = Color(0xFFCF2C27)
+val onErrorContainerLightMediumContrast = Color(0xFFFFFFFF)
+val backgroundLightMediumContrast = Color(0xFFF8F9FF)
+val onBackgroundLightMediumContrast = Color(0xFF191C20)
+val surfaceLightMediumContrast = Color(0xFFF5FAFB)
+val onSurfaceLightMediumContrast = Color(0xFF0C1213)
+val surfaceVariantLightMediumContrast = Color(0xFFDFE2EB)
+val onSurfaceVariantLightMediumContrast = Color(0xFF32363D)
+val outlineLightMediumContrast = Color(0xFF4F535A)
+val outlineVariantLightMediumContrast = Color(0xFF696D75)
+val scrimLightMediumContrast = Color(0xFF000000)
+val inverseSurfaceLightMediumContrast = Color(0xFF2B3133)
+val inverseOnSurfaceLightMediumContrast = Color(0xFFECF2F3)
+val inversePrimaryLightMediumContrast = Color(0xFFA3C9FE)
+val surfaceDimLightMediumContrast = Color(0xFFC2C7C9)
+val surfaceBrightLightMediumContrast = Color(0xFFF5FAFB)
+val surfaceContainerLowestLightMediumContrast = Color(0xFFFFFFFF)
+val surfaceContainerLowLightMediumContrast = Color(0xFFEFF5F6)
+val surfaceContainerLightMediumContrast = Color(0xFFE3E9EA)
+val surfaceContainerHighLightMediumContrast = Color(0xFFD8DEDF)
+val surfaceContainerHighestLightMediumContrast = Color(0xFFCDD3D4)
+
+val primaryLightHighContrast = Color(0xFF002D54)
+val onPrimaryLightHighContrast = Color(0xFFFFFFFF)
+val primaryContainerLightHighContrast = Color(0xFF214B78)
+val onPrimaryContainerLightHighContrast = Color(0xFFFFFFFF)
+val secondaryLightHighContrast = Color(0xFF4A2000)
+val onSecondaryLightHighContrast = Color(0xFFFFFFFF)
+val secondaryContainerLightHighContrast = Color(0xFF713B11)
+val onSecondaryContainerLightHighContrast = Color(0xFFFFFFFF)
+val tertiaryLightHighContrast = Color(0xFF452400)
+val onTertiaryLightHighContrast = Color(0xFFFFFFFF)
+val tertiaryContainerLightHighContrast = Color(0xFF6D3E04)
+val onTertiaryContainerLightHighContrast = Color(0xFFFFFFFF)
+val errorLightHighContrast = Color(0xFF600004)
+val onErrorLightHighContrast = Color(0xFFFFFFFF)
+val errorContainerLightHighContrast = Color(0xFF98000A)
+val onErrorContainerLightHighContrast = Color(0xFFFFFFFF)
+val backgroundLightHighContrast = Color(0xFFF8F9FF)
+val onBackgroundLightHighContrast = Color(0xFF191C20)
+val surfaceLightHighContrast = Color(0xFFF5FAFB)
+val onSurfaceLightHighContrast = Color(0xFF000000)
+val surfaceVariantLightHighContrast = Color(0xFFDFE2EB)
+val onSurfaceVariantLightHighContrast = Color(0xFF000000)
+val outlineLightHighContrast = Color(0xFF282C33)
+val outlineVariantLightHighContrast = Color(0xFF454951)
+val scrimLightHighContrast = Color(0xFF000000)
+val inverseSurfaceLightHighContrast = Color(0xFF2B3133)
+val inverseOnSurfaceLightHighContrast = Color(0xFFFFFFFF)
+val inversePrimaryLightHighContrast = Color(0xFFA3C9FE)
+val surfaceDimLightHighContrast = Color(0xFFB4BABB)
+val surfaceBrightLightHighContrast = Color(0xFFF5FAFB)
+val surfaceContainerLowestLightHighContrast = Color(0xFFFFFFFF)
+val surfaceContainerLowLightHighContrast = Color(0xFFECF2F3)
+val surfaceContainerLightHighContrast = Color(0xFFDEE3E5)
+val surfaceContainerHighLightHighContrast = Color(0xFFCFD5D6)
+val surfaceContainerHighestLightHighContrast = Color(0xFFC2C7C9)
+
+val primaryDark = Color(0xFFA3C9FE)
+val onPrimaryDark = Color(0xFF00315C)
+val primaryContainerDark = Color(0xFF1E4876)
+val onPrimaryContainerDark = Color(0xFFD3E4FF)
+val secondaryDark = Color(0xFFFFB688)
+val onSecondaryDark = Color(0xFF512400)
+val secondaryContainerDark = Color(0xFF6E380F)
+val onSecondaryContainerDark = Color(0xFFFFDBC7)
+val tertiaryDark = Color(0xFFFEB876)
+val onTertiaryDark = Color(0xFF4B2800)
+val tertiaryContainerDark = Color(0xFF6A3B02)
+val onTertiaryContainerDark = Color(0xFFFFDCBF)
+val errorDark = Color(0xFFFFB4AB)
+val onErrorDark = Color(0xFF690005)
+val errorContainerDark = Color(0xFF93000A)
+val onErrorContainerDark = Color(0xFFFFDAD6)
+val backgroundDark = Color(0xFF111418)
+val onBackgroundDark = Color(0xFFE1E2E8)
+val surfaceDark = Color(0xFF0E1415)
+val onSurfaceDark = Color(0xFFDEE3E5)
+val surfaceVariantDark = Color(0xFF43474E)
+val onSurfaceVariantDark = Color(0xFFC3C6CF)
+val outlineDark = Color(0xFF8D9199)
+val outlineVariantDark = Color(0xFF43474E)
+val scrimDark = Color(0xFF000000)
+val inverseSurfaceDark = Color(0xFFDEE3E5)
+val inverseOnSurfaceDark = Color(0xFF2B3133)
+val inversePrimaryDark = Color(0xFF39608F)
+val surfaceDimDark = Color(0xFF0E1415)
+val surfaceBrightDark = Color(0xFF343A3B)
+val surfaceContainerLowestDark = Color(0xFF090F10)
+val surfaceContainerLowDark = Color(0xFF171D1E)
+val surfaceContainerDark = Color(0xFF1B2122)
+val surfaceContainerHighDark = Color(0xFF252B2C)
+val surfaceContainerHighestDark = Color(0xFF303637)
+
+val primaryDarkMediumContrast = Color(0xFFC9DEFF)
+val onPrimaryDarkMediumContrast = Color(0xFF002649)
+val primaryContainerDarkMediumContrast = Color(0xFF6D93C5)
+val onPrimaryContainerDarkMediumContrast = Color(0xFF000000)
+val secondaryDarkMediumContrast = Color(0xFFFFD4BA)
+val onSecondaryDarkMediumContrast = Color(0xFF411B00)
+val secondaryContainerDarkMediumContrast = Color(0xFFC68051)
+val onSecondaryContainerDarkMediumContrast = Color(0xFF000000)
+val tertiaryDarkMediumContrast = Color(0xFFFFD4B0)
+val onTertiaryDarkMediumContrast = Color(0xFF3B1E00)
+val tertiaryContainerDarkMediumContrast = Color(0xFFC18347)
+val onTertiaryContainerDarkMediumContrast = Color(0xFF000000)
+val errorDarkMediumContrast = Color(0xFFFFD2CC)
+val onErrorDarkMediumContrast = Color(0xFF540003)
+val errorContainerDarkMediumContrast = Color(0xFFFF5449)
+val onErrorContainerDarkMediumContrast = Color(0xFF000000)
+val backgroundDarkMediumContrast = Color(0xFF111418)
+val onBackgroundDarkMediumContrast = Color(0xFFE1E2E8)
+val surfaceDarkMediumContrast = Color(0xFF0E1415)
+val onSurfaceDarkMediumContrast = Color(0xFFFFFFFF)
+val surfaceVariantDarkMediumContrast = Color(0xFF43474E)
+val onSurfaceVariantDarkMediumContrast = Color(0xFFD9DCE5)
+val outlineDarkMediumContrast = Color(0xFFAEB2BA)
+val outlineVariantDarkMediumContrast = Color(0xFF8D9099)
+val scrimDarkMediumContrast = Color(0xFF000000)
+val inverseSurfaceDarkMediumContrast = Color(0xFFDEE3E5)
+val inverseOnSurfaceDarkMediumContrast = Color(0xFF252B2C)
+val inversePrimaryDarkMediumContrast = Color(0xFF204977)
+val surfaceDimDarkMediumContrast = Color(0xFF0E1415)
+val surfaceBrightDarkMediumContrast = Color(0xFF3F4647)
+val surfaceContainerLowestDarkMediumContrast = Color(0xFF040809)
+val surfaceContainerLowDarkMediumContrast = Color(0xFF191F20)
+val surfaceContainerDarkMediumContrast = Color(0xFF23292A)
+val surfaceContainerHighDarkMediumContrast = Color(0xFF2D3435)
+val surfaceContainerHighestDarkMediumContrast = Color(0xFF393F40)
+
+val primaryDarkHighContrast = Color(0xFFE9F0FF)
+val onPrimaryDarkHighContrast = Color(0xFF000000)
+val primaryContainerDarkHighContrast = Color(0xFF9FC5FA)
+val onPrimaryContainerDarkHighContrast = Color(0xFF000C1D)
+val secondaryDarkHighContrast = Color(0xFFFFECE3)
+val onSecondaryDarkHighContrast = Color(0xFF000000)
+val secondaryContainerDarkHighContrast = Color(0xFFFFB17F)
+val onSecondaryContainerDarkHighContrast = Color(0xFF180600)
+val tertiaryDarkHighContrast = Color(0xFFFFEDDF)
+val onTertiaryDarkHighContrast = Color(0xFF000000)
+val tertiaryContainerDarkHighContrast = Color(0xFFF9B473)
+val onTertiaryContainerDarkHighContrast = Color(0xFF160800)
+val errorDarkHighContrast = Color(0xFFFFECE9)
+val onErrorDarkHighContrast = Color(0xFF000000)
+val errorContainerDarkHighContrast = Color(0xFFFFAEA4)
+val onErrorContainerDarkHighContrast = Color(0xFF220001)
+val backgroundDarkHighContrast = Color(0xFF111418)
+val onBackgroundDarkHighContrast = Color(0xFFE1E2E8)
+val surfaceDarkHighContrast = Color(0xFF0E1415)
+val onSurfaceDarkHighContrast = Color(0xFFFFFFFF)
+val surfaceVariantDarkHighContrast = Color(0xFF43474E)
+val onSurfaceVariantDarkHighContrast = Color(0xFFFFFFFF)
+val outlineDarkHighContrast = Color(0xFFEDF0F9)
+val outlineVariantDarkHighContrast = Color(0xFFBFC2CB)
+val scrimDarkHighContrast = Color(0xFF000000)
+val inverseSurfaceDarkHighContrast = Color(0xFFDEE3E5)
+val inverseOnSurfaceDarkHighContrast = Color(0xFF000000)
+val inversePrimaryDarkHighContrast = Color(0xFF204977)
+val surfaceDimDarkHighContrast = Color(0xFF0E1415)
+val surfaceBrightDarkHighContrast = Color(0xFF4B5152)
+val surfaceContainerLowestDarkHighContrast = Color(0xFF000000)
+val surfaceContainerLowDarkHighContrast = Color(0xFF1B2122)
+val surfaceContainerDarkHighContrast = Color(0xFF2B3133)
+val surfaceContainerHighDarkHighContrast = Color(0xFF363C3E)
+val surfaceContainerHighestDarkHighContrast = Color(0xFF424849)
+
