+++ core/designsystem/src/iosMain/kotlin/com/sedilant/yambol/core/designsystem/theme/PlatformTheme.ios.kt
@@ -1,0 +1,10 @@
+package com.sedilant.yambol.core.designsystem.theme
+
+import androidx.compose.material3.ColorScheme
+import androidx.compose.runtime.Composable
+
+@Composable
+actual fun platformDynamicColorScheme(darkTheme: Boolean): ColorScheme? {
+    // iOS does not support dynamic color schemes
+    return null
+}
