+++ core/domain/build.gradle.kts
@@ -1,9 +1,10 @@
 plugins {
     alias(libs.plugins.kotlin.multiplatform)
+    alias(libs.plugins.android.library)
 }
 
 kotlin {
-    jvm()
+    androidTarget()
     iosArm64()
     iosSimulatorArm64()
 
@@ -12,5 +13,19 @@ kotlin {
             implementation(libs.kotlinx.coroutines.core)
             implementation(libs.kotlinx.datetime)
         }
+    }
+}
+
+android {
+    namespace = "com.sedilant.yambol.core.domain"
+    compileSdk = 36
+
+    defaultConfig {
+        minSdk = 24
+    }
+
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
     }
 }
